fix(notifications): localize shared types
